pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Björn Petersen <git@bjoernpetersen.net>"]

[tool.poetry.dependencies]
python = "~3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
pydantic = "^1.9.0"
google-cloud-firestore = "^2.5.2"
pendulum = "^2.1.2"
python-telegram-bot = "^13.12"
requests = "^2.28.0"
more-itertools = "^8.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
